marketplace/requirements.txt
diploma-frontend==0.6
Django==4.2.1
django-taggit==4.0.0
djangorestframework==3.14.0
packaging==23.1
Pillow==9.5.0
